requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
PyGithub==2.3.0
python-dotenv==1.0.0
# ai-foundry-sdk (add actual package name when integrating AI Foundry)